requirements.txt
pip==24.0
asgiref==3.8.1
Django==5.2.1
sqlparse==0.5.3
djangorestframework~=3.16.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
sqlparse==0.5.3
